+++ src/cmd/mod.rs
@@ -1,6 +1,7 @@
 pub mod add;
 pub mod build;
 pub mod clean;
+pub mod doctor;
 pub mod import;
 pub mod init;
 pub mod jdk;